Revert "Snapshot search"
