--- snapshot/src/components/Home.js
@@ -1,10 +1,0 @@
-import React from 'react';
-
-function Home() {
-
-    return(
-        <h1>Home</h1>
-          
-      )
-}
-export default Home;
