+++ snapshot/src/components/Search.js
@@ -1,7 +1,9 @@
 import React from 'react';
 function Search(){
 return(
-    
+    <div>
+    <h1>Search</h1>
+    </div>
 )
 
 }
