+++ snapshot/src/index.css
@@ -16,9 +16,6 @@ code {
 form {
   width: 100%;
   padding: 20px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
 }
 
 .form-control {
@@ -30,6 +27,20 @@ form {
   border: 1px solid gray;
 }
 
+button[type="submit"] {
+  width: 20%;
+  background-color: #C37656;
+  color: white;
+  padding: 12px 20px;
+  border-radius: 4px;
+  border: none;
+  font-size: 1rem;
+  cursor: pointer;
+}
+
+button[type="submit"]:hover {
+  background-color: #bf6666;
+}
 nav{
   background-color: #F5EFEF;
 }
@@ -41,12 +52,4 @@ nav a {
   color: #1D639D;
   text-decoration: none;
 }
-.card {
-  width: 100%; 
-  height: 100%; 
-}
-.zoomed {
-  transform: scale(2.5);
-  z-index: 1;
-  position: relative;
-}
+
